Rename billing-monetize-subscriptions/urls to title
